All re-writes for cumulative-monthly improvements
